+++ libraries/transformer/src/test/java/androidx/media3/transformer/AudioGraphInputTest.java
@@ -18,7 +18,10 @@ package androidx.media3.transformer;
 
 import static androidx.media3.common.util.Assertions.checkState;
 import static androidx.media3.common.util.Util.getPcmFormat;
+import static androidx.media3.transformer.TestUtil.createSpeedChangingAudioProcessor;
 import static com.google.common.truth.Truth.assertThat;
+import static java.util.Collections.max;
+import static java.util.Collections.min;
 
 import androidx.media3.common.C;
 import androidx.media3.common.Format;
@@ -28,6 +31,7 @@ import androidx.media3.common.util.Util;
 import androidx.media3.decoder.DecoderInputBuffer;
 import androidx.media3.test.utils.TestUtil;
 import androidx.test.ext.junit.runners.AndroidJUnit4;
+import com.google.common.collect.ImmutableList;
 import com.google.common.primitives.Bytes;
 import java.nio.ByteBuffer;
 import java.util.ArrayList;
@@ -40,6 +44,12 @@ import org.junit.runner.RunWith;
 public class AudioGraphInputTest {
   private static final EditedMediaItem FAKE_ITEM =
       new EditedMediaItem.Builder(MediaItem.EMPTY).build();
+  private static final EditedMediaItem FAKE_ITEM_WITH_DOUBLE_SPEED =
+      new EditedMediaItem.Builder(MediaItem.EMPTY)
+          .setEffects(
+              new Effects(
+                  ImmutableList.of(createSpeedChangingAudioProcessor(2)), ImmutableList.of()))
+          .build();
   private static final AudioFormat MONO_44100 =
       new AudioFormat(/* sampleRate= */ 44_100, /* channelCount= */ 1, C.ENCODING_PCM_16BIT);
   private static final AudioFormat MONO_48000 =
@@ -250,10 +260,10 @@ public class AudioGraphInputTest {
             /* editedMediaItem= */ FAKE_ITEM,
             /* inputFormat= */ getPcmFormat(STEREO_44100));
     byte[] inputData = TestUtil.buildTestData(/* length= */ 100 * STEREO_44100.bytesPerFrame);
-
+    // Pass in duration approximately equal to raw data duration ~ 100 / 44100 ~ 2267us.
     audioGraphInput.onMediaItemChanged(
         /* editedMediaItem= */ FAKE_ITEM,
-        /* durationUs= */ 1_000_000,
+        /* durationUs= */ 2267,
         /* decodedFormat= */ getPcmFormat(STEREO_44100),
         /* isLast= */ true);
 
@@ -275,6 +285,90 @@ public class AudioGraphInputTest {
   }
 
   @Test
+  public void getOutput_withNoEffects_returnsInputDataAndSilence() throws Exception {
+    AudioGraphInput audioGraphInput =
+        new AudioGraphInput(
+            /* requestedOutputAudioFormat= */ AudioFormat.NOT_SET,
+            /* editedMediaItem= */ FAKE_ITEM,
+            /* inputFormat= */ getPcmFormat(STEREO_44100));
+    byte[] inputData = TestUtil.buildTestData(/* length= */ 100 * STEREO_44100.bytesPerFrame);
+
+    audioGraphInput.onMediaItemChanged(
+        /* editedMediaItem= */ FAKE_ITEM,
+        /* durationUs= */ 1_000_000,
+        /* decodedFormat= */ getPcmFormat(STEREO_44100),
+        /* isLast= */ true);
+
+    // Force the media item change to be processed.
+    checkState(!audioGraphInput.getOutput().hasRemaining());
+
+    // Queue inputData: 100 * STEREO_44100.bytesPerFrame bytes = 100 PCM samples.
+    // Audio duration is 100 / 44100 seconds ~ 2_268us.
+    DecoderInputBuffer inputBuffer = audioGraphInput.getInputBuffer();
+    inputBuffer.ensureSpaceForWrite(inputData.length);
+    inputBuffer.data.put(inputData).flip();
+    checkState(audioGraphInput.queueInputBuffer());
+
+    // Queue EOS. Input audio track ends before onMediaItemChanged durationUs = 1_000_000.
+    // AudioGraphInput will append generated silence up to target durationUs of 1s (~997_732us).
+    audioGraphInput.getInputBuffer().setFlags(C.BUFFER_FLAG_END_OF_STREAM);
+    checkState(audioGraphInput.queueInputBuffer());
+
+    List<Byte> outputBytes = drainAudioGraphInputUntilEnded(audioGraphInput);
+    long expectedSampleCount = Util.durationUsToSampleCount(1_000_000, STEREO_44100.sampleRate);
+    // Silent audio generator rounds up duration.
+    assertThat(outputBytes.size())
+        .isEqualTo((expectedSampleCount + 1) * STEREO_44100.bytesPerFrame);
+    assertThat(outputBytes.subList(0, inputData.length))
+        .containsExactlyElementsIn(Bytes.asList(inputData))
+        .inOrder();
+    assertThat(min(outputBytes.subList(inputData.length, outputBytes.size()))).isEqualTo(0);
+    assertThat(max(outputBytes.subList(inputData.length, outputBytes.size()))).isEqualTo(0);
+  }
+
+  @Test
+  public void getOutput_withEffects_returnsInputDataAndSilence() throws Exception {
+    AudioGraphInput audioGraphInput =
+        new AudioGraphInput(
+            /* requestedOutputAudioFormat= */ AudioFormat.NOT_SET,
+            /* editedMediaItem= */ FAKE_ITEM_WITH_DOUBLE_SPEED,
+            /* inputFormat= */ getPcmFormat(STEREO_44100));
+    byte[] inputData = TestUtil.buildTestData(/* length= */ 4096 * STEREO_44100.bytesPerFrame);
+
+    audioGraphInput.onMediaItemChanged(
+        /* editedMediaItem= */ FAKE_ITEM_WITH_DOUBLE_SPEED,
+        /* durationUs= */ 1_000_000,
+        /* decodedFormat= */ getPcmFormat(STEREO_44100),
+        /* isLast= */ true);
+
+    // Force the media item change to be processed.
+    checkState(!audioGraphInput.getOutput().hasRemaining());
+
+    // Queue inputData: 4096 * STEREO_44100.bytesPerFrame bytes = 100 PCM samples.
+    // Audio duration is 4096 / 44100 seconds ~ 92_880us.
+    DecoderInputBuffer inputBuffer = audioGraphInput.getInputBuffer();
+    inputBuffer.ensureSpaceForWrite(inputData.length);
+    inputBuffer.data.put(inputData).flip();
+    checkState(audioGraphInput.queueInputBuffer());
+
+    // Queue EOS. Input audio track ends before onMediaItemChanged durationUs = 1_000_000.
+    // AudioGraphInput will append generated silence up to target durationUs of 1s (~907_120us).
+    audioGraphInput.getInputBuffer().setFlags(C.BUFFER_FLAG_END_OF_STREAM);
+    checkState(audioGraphInput.queueInputBuffer());
+
+    List<Byte> outputBytes = drainAudioGraphInputUntilEnded(audioGraphInput);
+    long expectedSampleCount = Util.durationUsToSampleCount(500_000, STEREO_44100.sampleRate);
+    // Silent audio generator rounds up duration.
+    assertThat(outputBytes.size())
+        .isEqualTo((expectedSampleCount + 1) * STEREO_44100.bytesPerFrame);
+    // Sonic takes a while to zero-out the input.
+    assertThat(min(outputBytes.subList(inputData.length * 6 / 10, outputBytes.size())))
+        .isEqualTo(0);
+    assertThat(max(outputBytes.subList(inputData.length * 6 / 10, outputBytes.size())))
+        .isEqualTo(0);
+  }
+
+  @Test
   public void getOutput_withSilentMediaItemChange_outputsCorrectAmountOfSilentBytes()
       throws Exception {
     AudioGraphInput audioGraphInput =
@@ -291,6 +385,26 @@ public class AudioGraphInputTest {
 
     int bytesOutput = drainAudioGraphInputUntilEnded(audioGraphInput).size();
     long expectedSampleCount = Util.durationUsToSampleCount(1_000_000, STEREO_44100.sampleRate);
+    assertThat(bytesOutput).isEqualTo(expectedSampleCount * STEREO_44100.bytesPerFrame);
+  }
+
+  @Test
+  public void getOutput_withSilentMediaItemAndEffectsChange_outputsCorrectAmountOfSilentBytes()
+      throws Exception {
+    AudioGraphInput audioGraphInput =
+        new AudioGraphInput(
+            /* requestedOutputAudioFormat= */ AudioFormat.NOT_SET,
+            /* editedMediaItem= */ FAKE_ITEM_WITH_DOUBLE_SPEED,
+            /* inputFormat= */ getPcmFormat(STEREO_44100));
+
+    audioGraphInput.onMediaItemChanged(
+        /* editedMediaItem= */ FAKE_ITEM_WITH_DOUBLE_SPEED,
+        /* durationUs= */ 1_000_000,
+        /* decodedFormat= */ null,
+        /* isLast= */ true);
+
+    int bytesOutput = drainAudioGraphInputUntilEnded(audioGraphInput).size();
+    long expectedSampleCount = Util.durationUsToSampleCount(500_000, STEREO_44100.sampleRate);
     assertThat(bytesOutput).isEqualTo(expectedSampleCount * STEREO_44100.bytesPerFrame);
   }